# perf: embeddings