Benchmark


```
docs = [Document(embedding=np.array([1,2,3]),
                 text=f'{i}') for i in range(100_000)]

da = DocumentArray(docs)
```

- On master
          ```
          %timeit aux = da.embeddings     1.22 s ± 30.7 ms
          ```
- This PR
          ```
          %timeit aux = da.embeddings      80.2 ms ± 433 µs
          ```


This has a huge impact on `.match`

```
da2 = DocumentArray([Document(embedding=np.array([3,4,5]))])
```

- On master
          ```
          %timeit aux = da.embeddings          1.13 s ± 16.8 ms 
          ```
- This PR
          ```
          %timeit aux = da.embeddings        77.3 ms ± 1.13 ms 
          ```